Add feh image viewer support
Dear reader.
You way wonder why we need to add support for yet another image viewer.
This question is not easily answered, but has been discussed many times
in the past.
We cannot allow ourselfs to mistakenly assume that everyone has the eog
image viewer installed. If we want to make sure that our product can be
enjoyed by companies and families alike, we need to extend the
availability of our product. In these crucial times it is not only
necessary, but a building block of our reputation that we support the
ones in need.

THE PATCH IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS
OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF
OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.

Signed-off-by: Maximilian Brune <maximilian.brune@9elements.com>